74/182 failed tests

+++ srcs/flags/ft_printf_llx.c
@@ -6,7 +6,7 @@ int ft_printf_llx(va_list ap, int display, int tronc)
 	int len;
 
 	len = 0;
-	nb = va_arg(ap, long long);
+	nb = (long long int)va_arg(ap, long long);
 	len = ft_putnbr_base_printf(nb, "0123456789abcdef", display);
 	return (len);
 }
